requirements.txt
alembic==1.10.4
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
black==23.3.0
click==8.1.3
databases==0.6.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.1
fastapi-pagination==0.12.3
FastAPI-SQLAlchemy==0.2.1
flake8==6.0.0
h11==0.14.0
idna==3.4
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.5.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
PyJWT==2.7.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.41.1
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.22.0
